requirements.txt
annotated-types==0.7.0
anthropic==0.30.1
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
filelock==3.15.4
FLAML==2.1.2
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
jiter==0.5.0
numpy==1.26.4
openai==1.35.9
packaging==24.1
pyautogen==0.2.31
pydantic==2.8.0
pydantic_core==2.20.0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
sniffio==1.3.1
termcolor==2.4.0
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
